smart_rag_agent/agents/core.py


from openai import OpenAI
client = OpenAI()

# 1. ModelSettings — just a tiny container
class ModelSettings:
    def __init__(self, max_tokens=512):
        self.max_tokens = max_tokens

# 2. @function_tool — does nothing fancy, just marks the function
def function_tool(func):
    return func

# 3. SQLiteSession — only needed to exist (your code never really uses it deeply)
class SQLiteSession:
    def __init__(self, name): pass
    def close(self): pass

# 4. Agent — just stores the settings (never used beyond this)
class Agent:
    def __init__(self, name, instructions, tools=None, handoffs=None, model="gpt-4o-mini", model_settings=None):
        self.name = name
        self.instructions = instructions
        self.tools = tools or []
        self.model = model
        self.model_settings = model_settings or ModelSettings()

# 5. Runner — the ONLY thing that really matters
class Runner:
    @staticmethod
    def run_sync(agent, user_input, session=None, max_turns=3):
        # Find the retrieval tool (it's always the first one in your code)
        retrieval_tool = next((t for t in agent.tools if t.__name__ == "retrieve_context"), None)

        # First ask GPT
        messages = [
            {"role": "system", "content": agent.instructions},
            {"role": "user", "content": user_input}
        ]

        response = client.chat.completions.create(
            model=agent.model,
            messages=messages,
            max_tokens=agent.model_settings.max_tokens
        )
        answer = response.choices[0].message.content

        # If we have context tool and answer looks weak → force context
        if retrieval_tool and ("NO_CONTEXT_FOUND" in answer or len(answer) < 50):
            context = retrieval_tool(user_input)  # call your retrieve_context
            messages.append({"role": "assistant", "content": answer})
            messages.append({"role": "user", "content": f"Context:\n{context}\n\nNow answer properly."})

            final = client.chat.completions.create(
                model=agent.model,
                messages=messages,
                max_tokens=agent.model_settings.max_tokens
            )
            answer = final.choices[0].message.content

        # Fake result object — your code only reads .final_output
        class Result:
            final_output = answer.strip()

        return Result()