smart_rag_agent/requirements.txt
streamlit
requests
httpx
beautifulsoup4
chromadb
langchain-openai
langchain-text-splitters
nest_asyncio
python-dotenv
openai-agents